# perf: Optimize isRecursive check in DynamicModelCompiler