@kubukoz 2024-04-04
Probably a good idea to run some profiling first as well, it's possible there are other bottlenecks in there