@kubukoz 2024-04-11
I made a separate ticket for the codegen part so that it's easier to track and discuss on its own: https://github.com/disneystreaming/smithy4s/issues/1483